ToDo/requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.6
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djoser==2.0.3
kombu==4.6.8
PyJWT==1.7.1
pytz==2020.1
redis==3.5.2
sqlparse==0.3.1
vine==1.3.0
